Reorder the pom
This order confirms to the maven developer's team conventions. [1]
This also adds and xml declaration and reformats the root element,
which came as a result of me using mvn tidy:pom,
but then deciding to use multiple lines instead of one long line.
This doesn't conform to the conventions below, but I'm ok with that.

[1]: https://maven.apache.org/developers/conventions/code.html#POM_Code_Convention